+++ activesupport/lib/active_support/core_ext/object/conversions.rb
@@ -2,5 +2,6 @@
 
 require "active_support/core_ext/object/to_param"
 require "active_support/core_ext/object/to_query"
+require "active_support/core_ext/object/as_tags"
 require "active_support/core_ext/array/conversions"
 require "active_support/core_ext/hash/conversions"